Codeship: Added solution for Count Commas In Range Ii

+++ problems/count-commas-in-range-ii/solution.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long countCommas(long long n) {
         int ans = 0;
-        int x = 1000;
+        long long x = 1000;
         while(x <= n){
             ans += n - x + 1;
             x *= 1000;
